requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.1.7
django-autocomplete-light==3.3.2
django-ckeditor==5.6.1
django-debug-toolbar==1.11
django-encrypted-model-fields==0.5.8
django-environ==0.4.5
django-filter==2.1.0
django-js-asset==1.2.2
-e git+https://github.com/ninapavlich/django-list-wrestler.git@0ecd650587326cb1dfb9fc4c79fd9e9f3e881762#egg=django_list_wrestler
djangorestframework==3.9.2
idna==2.8
lorem==0.1.1
Markdown==3.0.1
pycparser==2.19
python-magic==0.4.15
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
Unidecode==1.0.23
urllib3==1.24.1
